requirements.txt
python==3.10.14
pandas==2.2.2 
pillow==10.3.0
numpy==1.26.4 
pylint==3.2.3
matplotlib==3.9.0 
tqdm==4.66.4 
segmentation-models-pytorch==0.3.3  
pytorch-lightning==2.3.0 
gputil==1.4.0 